Add GameLoop.java for smoother movement

+++ src/main/java/com/geowars/core/util/Constants.java
@@ -1,6 +1,6 @@
 package com.geowars.core.util;
 
 public class Constants {
-    public static final int WINDOW_WIDTH = 1024;
-    public static final int WINDOW_HEIGHT = 768;
+    public static final double WINDOW_WIDTH = 1024;
+    public static final double WINDOW_HEIGHT = 768;
 }
